Always use TLS MITM for CONNECT if possible.
CONNECT is usually used by clients when the target is over HTTPs. As
long as we've a CA key pair specified, the CONNECT handler should do TLS
MITM.

The target port doesn't actually matter. We assume that clients
use the right method to talk with the proxy. In other words, we assume
that CONNECT is not used when forwarding to a plain HTTP target. This
use-case should be very minimal / non-existent.